删除没用的文件；修改代码格式；修改编译指令，忽略memset()函数的警告；

+++ Link22/Utils/CrcUtil.cpp
@@ -61,7 +61,7 @@ bitset<16> CrcUtil::crc16(uint16 data[], int len) {
     uint CRC_SIZE16 = 16;
     uint remainder = 0;
     int nremainder = 0;
-    int j,k;
+    size_t j,k;
     uint bit,lb;
     uint fb,cb;
     for (int i = 0; i < len; i++) {
